Refactor duplicate bkey processing logic into its own fuction.
This also changes the logic to switch on KEY_TYPE instead of using the bkey_is_x() functions.
Most notable reason for this is that the KEY_TYPE_btree_ptr(_v2) type was causing unexpected behaviour.
These are matched as part of bkey_extent_is_data() but don't seem to point to file contents.